third_party/blink/web_tests/fast/js/regress/resolved-timezone-defined-expected.txt

PASS new Intl.DateTimeFormat('en').resolvedOptions().timeZone is defined.
PASS successfullyParsed is true

TEST COMPLETE
